Bug 77992 part 3 - Make event.timeStamp return a DOMHighResTimeStamp; r=bz
Even when dom.event.highrestimestamp.enabled is false and we should return
a DOMTimeStamp (=unsigned long long), we still just return a double since that's
what would happen anyway when we convert the DOMTimeStamp to a Javascript
Number and it simplifies the code significantly to just always return a double.

+++ dom/events/Event.h
@@ -193,7 +193,7 @@ public:
     return mEvent->mFlags.mIsSynthesizedForTests;
   }
 
-  uint64_t TimeStamp() const;
+  double TimeStamp() const;
 
   void InitEvent(const nsAString& aType, bool aBubbles, bool aCancelable,
                  ErrorResult& aRv)
